@@ -1,130 +1,88 @@
 {
   "name": "identity-provider",
   "version": "1.0.4",
-  "repository": "https://gitlab.com/castlecraft/building-blocks",
-  "description": "Identity Provider",
-  "author": "CEPL",
-  "license": "AGPLv3",
-  "private": true,
   "scripts": {
     "ng": "ng",
-    "start:client": "ng serve --proxy-config proxy.config.json --open",
-    "lint": "tslint -c tslint.json 'src/**/*.ts'",
-    "e2e": "echo ng e2e --port 49100 --configuration=ci",
     "format": "prettier --write \"**/*.ts\"",
-    "format:check": "prettier-check \"**/*.ts\"",
-    "start": "nodemon",
-    "start:hmr": "node dist/server",
-    "prestart:prod": "rm -rf dist && ng build --prod && ng run identity-provider:server:production && tsc -p src/server/tsconfig.server.json",
-    "start:prod": "export NODE_ENV=production && node dist/out-tsc/main.js",
-    "start:server": "ts-node -r tsconfig-paths/register --project src/server/tsconfig.server.json src/server/main.ts",
-    "test": "export NODE_ENV=test && jest",
-    "test:client": "echo ng test --watch=false --no-progress",
-    "test:cov": "jest --coverage",
-    "test:e2e": "export NODE_ENV=test && jest --config ./tests/jest-e2e.json",
+    "ng:start": "ng serve",
+    "e2e": "ng e2e --configuration=ci",
     "build": "ng build",
-    "build:client-and-server-bundles": "ng build --prod && ng run identity-provider:server:production",
-    "build:prerender": "npm run build:client-and-server-bundles && npm run webpack:server && npm run generate:prerender",
-    "build:ssr": "npm run build:client-and-server-bundles && npm run webpack:server",
-    "generate:prerender": "cd dist && node prerender",
-    "webpack:server": "webpack --config webpack.server.config.js --progress --colors",
-    "serve:prerender": "cd dist/identity-provider && http-server",
-    "serve:ssr": "node dist/server",
-    "start:dev": "concurrently --kill-others-on-fail \"npm run start\" \"npm run start:client\"",
-    "add:user": "ts-node -r tsconfig-paths/register --project src/server/tsconfig.server.json src/server/cli-connector/create-user.ts",
-    "add:oauth2-client": "ts-node -r tsconfig-paths/register --project src/server/tsconfig.server.json src/server/cli-connector/create-client.ts"
+    "lint": "./node_modules/.bin/tslint -c tslint.json 'src/**/*.ts'",
+    "start:server": "nodemon",
+    "start:client": "ng serve --proxy-config proxy.config.json",
+    "start:dev": "concurrently --kill-others-on-fail \"yarn start:server\" \"yarn start:client\"",
+    "prestart:prod": "rm -rf dist && tsc",
+    "start:prod": "node dist/main.js",
+    "start:hmr": "node dist/server",
+    "test:server": "jest",
+    "test:cov": "jest --coverage",
+    "test:e2e": "jest --config ./tests/jest-e2e.json --detectOpenHandles",
+    "test:client": "ng test --no-progress",
+    "webpack": "webpack --config webpack.config.js"
   },
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
-    "@angular/cdk": "^6.4.1",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.7",
-    "@angular/flex-layout": "^6.0.0-beta.17",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.4.1",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/platform-server": "^6.1.0",
-    "@angular/router": "^6.0.3",
-    "@nest-middlewares/passport": "^5.0.2",
-    "@nestjs/common": "^5.1.0",
-    "@nestjs/core": "^5.1.0",
-    "@nestjs/microservices": "^5.1.0",
-    "@nestjs/passport": "^1.1.0",
-    "@nestjs/swagger": "^2.2.0",
-    "@nestjs/typeorm": "^5.0.0",
-    "@nestjs/websockets": "^5.1.0",
-    "@nguniversal/express-engine": "^6.0.0",
-    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
-    "@types/passport-local": "^1.0.33",
+    "@angular/animations": "^6.1.6",
+    "@angular/common": "^6.1.6",
+    "@angular/compiler": "^6.1.6",
+    "@angular/core": "^6.1.6",
+    "@angular/forms": "^6.1.6",
+    "@angular/http": "^6.1.6",
+    "@angular/platform-browser": "^6.1.6",
+    "@angular/platform-browser-dynamic": "^6.1.6",
+    "@angular/router": "^6.1.6",
+    "@nestjs/common": "^5.3.5",
+    "@nestjs/core": "^5.3.4",
+    "@nestjs/swagger": "^2.5.1",
+    "@nestjs/typeorm": "^5.2.0",
     "class-transformer": "^0.1.9",
-    "class-validator": "^0.8.5",
-    "cookie-parser": "^1.4.3",
+    "class-validator": "^0.9.1",
     "core-js": "^2.5.4",
-    "dotenv": "^6.0.0",
-    "express-session": "^1.15.6",
-    "hammerjs": "^2.0.8",
-    "hbs": "^4.0.1",
-    "mysql": "^2.15.0",
-    "nestjs-ensureloggedin-guard": "^1.0.2",
-    "nestjs-session-store": "^1.0.3",
-    "oauth2orize": "^1.11.0",
-    "passport": "^0.4.0",
-    "passport-http": "^0.3.0",
-    "passport-http-bearer": "^1.0.1",
-    "passport-local": "^1.0.0",
-    "passport-oauth2-client-password": "^0.1.2",
-    "passport-oauth2-code": "^0.1.6",
-    "point-of-view": "^1.1.0",
+    "fastify-formbody": "^2.0.0",
+    "joi": "^13.6.0",
+    "mongodb": "~3.0.11",
     "reflect-metadata": "^0.1.12",
-    "rxjs": "^6.0.0",
+    "rxjs": "^6.3.2",
     "typeorm": "^0.2.7",
-    "yargs": "^12.0.1",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
-    "@nestjs/testing": "^5.1.0",
+    "@angular-devkit/build-angular": "~0.7.5",
+    "@angular/cli": "~6.1.5",
+    "@angular/compiler-cli": "^6.1.6",
+    "@angular/language-service": "^6.1.6",
+    "@nestjs/testing": "^5.3.1",
     "@types/express": "^4.0.39",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
-    "@types/jest": "23.1.4",
-    "@types/joi": "^13.3.0",
-    "@types/node": "10.5.2",
+    "@types/jest": "^23.3.1",
+    "@types/node": "^10.9.4",
     "@types/supertest": "^2.0.4",
-    "@types/yargs": "^11.1.1",
-    "codelyzer": "^4.4.2",
-    "concurrently": "^3.6.0",
-    "jasmine-core": "^3.1.0",
+    "codelyzer": "~4.4.4",
+    "concurrently": "^4.0.1",
+    "jasmine-core": "~3.2.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^23.2.0",
-    "karma": "^2.0.4",
+    "jest": "^23.5.0",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-coverage-istanbul-reporter": "~2.0.3",
     "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^1.2.0",
-    "nodemon": "^1.14.1",
+    "karma-jasmine-html-reporter": "^1.3.1",
+    "nodemon": "^1.18.4",
     "prettier": "^1.11.1",
-    "prettier-check": "^2.0.0",
-    "protractor": "^5.4.0",
-    "supertest": "^3.0.0",
-    "ts-jest": "^23.0.0",
+    "protractor": "~5.4.0",
+    "supertest": "^3.2.0",
+    "ts-jest": "^23.1.4",
     "ts-loader": "^4.1.0",
-    "ts-node": "7.0.0",
+    "ts-node": "^7.0.1",
     "tsconfig-paths": "^3.1.1",
-    "tslint": "^5.10.0",
+    "tslint": "~5.11.0",
     "typescript": "~2.7.2",
-    "webpack": "4.15.1",
-    "webpack-cli": "3.0.8",
+    "webpack": "^4.17.2",
+    "webpack-cli": "^3.1.0",
     "webpack-node-externals": "^1.6.0"
   },
   "jest": {
-    "testEnvironment": "node",
     "moduleDirectories": [
       "src/server",
       "node_modules"
