@@ -1,32 +1,17 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HttpClientModule } from '@angular/common/http';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { AuthServerMaterialModule } from './auth-server-material/auth-server-material.module';
-import { FormsModule } from '@angular/forms';
-import { LoginComponent } from './login/login.component';
-import { SignupComponent } from './signup/signup.component';
-import { AccountComponent } from './account/account.component';
+import { AppService } from './app.service';
+import { HttpErrorHandler } from './http-error-handler.service';
+import { MessageService } from './message.service';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    SignupComponent,
-    AccountComponent,
-  ],
-  imports: [
-    BrowserModule.withServerTransition({ appId: 'identity-provider' }),
-    AppRoutingModule,
-    HttpClientModule,
-    BrowserAnimationsModule,
-    AuthServerMaterialModule,
-    FormsModule,
-  ],
-  providers: [],
+  declarations: [AppComponent],
+  imports: [BrowserModule, AppRoutingModule, HttpClientModule],
+  providers: [AppService, HttpErrorHandler, MessageService],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
