@@ -27,10 +27,11 @@ describe('AppController (e2e)', () => {
     await app.init();
   });
 
-  it('/GET /message', () => {
+  it('/GET /message', (done) => {
     return request(app.getHttpServer())
       .get('/message')
       .expect(200)
-      .expect({ message: 'NestJS' });
+      .expect({ message: 'NestJS' })
+      .end(done);
   });
 });
