@@ -1,25 +1,10 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule, CanActivate } from '@angular/router';
-import { AccountComponent } from './account/account.component';
-import { LoginComponent } from './login/login.component';
-import { SignupComponent } from './signup/signup.component';
-import { AuthGuardService } from './auth/auth-guard.service';
-import { AuthService } from './auth/auth.service';
+import { Routes, RouterModule } from '@angular/router';
 
-const routes: Routes = [
-  { path: '', redirectTo: 'login', pathMatch: 'full' },
-  { path: 'login', component: LoginComponent },
-  {
-    path: 'account',
-    component: AccountComponent,
-    canActivate: [AuthGuardService],
-  },
-  { path: 'signup', component: SignupComponent },
-];
+const routes: Routes = [];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
   exports: [RouterModule],
-  providers: [AuthService, AuthGuardService],
 })
 export class AppRoutingModule {}
