@@ -1,11 +1,25 @@
 import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
+import { AppService } from './app.service';
+import { from } from 'rxjs';
+
 describe('AppComponent', () => {
+  const appServiceStub: Partial<AppService> = {
+    getMessage: () => {
+      return from([]);
+    },
+  };
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [RouterTestingModule],
       declarations: [AppComponent],
+      providers: [
+        {
+          provide: AppService,
+          useValue: appServiceStub,
+        },
+      ],
     }).compileComponents();
   }));
   it('should create the app', async(() => {
@@ -13,9 +27,9 @@ describe('AppComponent', () => {
     const app = fixture.debugElement.componentInstance;
     expect(app).toBeTruthy();
   }));
-  it(`should have as title 'identity-provider'`, async(() => {
+  it(`should have as title 'app'`, async(() => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('identity-provider');
+    expect(app).toBeDefined();
   }));
 });
