@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 @Injectable()
 export class AppService {
-  root(): string {
-    return 'IDP built using Nest.js';
+  message() {
+    return { message: 'NestJS' };
   }
 }
