@@ -1,12 +1,43 @@
+import * as Joi from 'joi';
 import * as fs from 'fs';
+import * as dotenv from 'dotenv';
+
+export interface EnvConfig {
+  [prop: string]: string;
+}
 
 export class ConfigService {
-  getConfig(configName) {
-    return JSON.parse(
-      fs.readFileSync(
-        `config/${configName}/${process.env.NODE_ENV}.json`,
-        'utf-8',
-      ),
+  private readonly envConfig: EnvConfig;
+
+  constructor() {
+    const config = dotenv.config().parsed;
+    this.envConfig = this.validateInput(config);
+  }
+
+  /**
+   * Ensures all needed variables are set, and returns the validated JavaScript object
+   * including the applied default values.
+   */
+  private validateInput(envConfig: EnvConfig): EnvConfig {
+    const envVarsSchema: Joi.ObjectSchema = Joi.object({
+      NODE_ENV: Joi.string()
+        .valid(['development', 'production', 'test', 'provision'])
+        .default('development'),
+      DB_NAME: Joi.string().required(),
+      DB_HOST: Joi.string().required(),
+    });
+
+    const { error, value: validatedEnvConfig } = Joi.validate(
+      envConfig,
+      envVarsSchema,
     );
+    if (error) {
+      throw new Error(`Config validation error: ${error.message}`);
+    }
+    return validatedEnvConfig;
+  }
+
+  get(key: string): string {
+    return this.envConfig[key];
   }
 }
