@@ -1,29 +1,15 @@
 import { ConfigService } from '../config/config.service';
-import { AuthData } from './auth-data/auth-data.entity';
-import { AuthorizationCode } from './authorization-code/authorization-code.entity';
-import { BearerToken } from './bearer-token/bearer-token.entity';
-import { Client } from './client/client.entity';
-import { Role } from './role/role.entity';
-import { Session } from './session/session.entity';
-import { User } from './user/user.entity';
+import { IdentityProviderSettings } from './identity-provider-settings/identity-provider-settings.entity';
+import { MongoConnectionOptions } from 'typeorm/driver/mongodb/MongoConnectionOptions';
+import { Profile } from './profile/profile.entity';
 
 const config = new ConfigService();
-const typeormConnection = config.getConfig('ormconfig');
-let defaultConnection;
-typeormConnection.forEach(element => {
-  if (element.name === 'default') {
-    defaultConnection = element;
-  }
-});
 
-defaultConnection.entities = [
-  AuthData,
-  AuthorizationCode,
-  BearerToken,
-  Client,
-  Role,
-  Session,
-  User,
-];
-
-export const TYPEORM_CONNECTION = defaultConnection;
+export const TYPEORM_CONNECTION: MongoConnectionOptions = {
+  type: 'mongodb',
+  host: config.get('DB_HOST'),
+  database: config.get('DB_NAME'),
+  logging: false,
+  synchronize: true,
+  entities: [IdentityProviderSettings, Profile],
+};
